Add some testing for uvcal select method
